screen_recorder/requirements.txt
customtkinter
mss
opencv-python
sounddevice
numpy
keyboard
pystray
Pillow
